Script now working but not using peep

+++ installBedrock.sh
@@ -16,7 +16,7 @@ function echogreen() {
 }
 
 echogreen "BEDROCK installation script on Ubuntu"
-echo "Warning: This install script works on Ubuntu 13.04 (64bits)"
+echo "Warning: This install script works on Ubuntu 14.10 (64bits)"
 echo "Warning: There is no error handling whatsoever."
 echo "Feel free to fork it and adapt it for another distro or update your changes"
 echo "This script is going to install Bedrock in a bedrock folder."
@@ -75,14 +75,16 @@ else
     fi
 fi
 
+echogreen "Get latest commits from upstream Bedrock"
+git pull upstream master
+git submodule update --init --recursive
+
 echogreen "Create a virtual environement in the folder venv"
 virtualenv venv             # create a virtual env in the folder `venv`
 echo "Activate the virtual env"
 source ./venv/bin/activate    # activate the virtual env
 echogreen "Install Bedrock local dependencies in venv"
-chmod 755 ./bin/peep.py
-./bin/peep.py install -r requirements/compiled.txt --no-use-wheel
-./bin/peep.py install -r requirements/prod.txt --no-use-wheel
+./venv/bin/pip install -r requirements/prod.txt
 
 echored "Do you want to install developper dependencies to be able to run tests locally and participate to documentation? (y/n)"
 read -n 1 devdependencies
@@ -90,28 +92,36 @@ echo ""
 if [ $devdependencies == 'y' ]
 then
     echogreen "Installing developer dependencies..."
-    ./bin/peep.py install -r requirements/dev.txt --no-use-wheel
+    ./venv/bin/pip install -r requirements/dev.txt
 fi
 
 ./venv/bin/pip install ipython        # highly recommended, but not required so not in requirements/dev.txt
 
-echogreen "npm install: less, grunt-cli, jshint"
-sudo npm install -g less
-sudo npm install -g grunt-cli
-sudo npm install -g jshint
-sudo npm install
-#echo -e "\nLESS_BIN = '/usr/bin/lessc'" >> bedrock/settings/local.py
+echored "Do you want to install npm dependencies globally (sudo needed)? (y/n)"
+read -n 1 npmdependencies
+echo ""
+if [ $npmdependencies == 'y' ]
+then
+    echogreen "npm install: less, grunt-cli, jshint"
+    sudo npm install -g less
+    sudo npm install -g grunt-cli
+    sudo npm install -g jshint
+    sudo npm install
+fi
 
 echogreen "Copy bedrock/settings/local.py-dist into bedrock/settings/local.py"
 cp bedrock/settings/local.py-dist bedrock/settings/local.py
+sed -i -- "s/STATSD_HOST = ''/STATSD_HOST = 'localhost'/g" bedrock/settings/local.py
 
-echogreen "Sync database schemas"
-# mysql-ctl start
-./bin/sync_all
+find . -name '*.pyc' -exec rm {} \;
 
 echogreen "Check out all the translations which live on svn in the localizers repositories"
+# ln -s ~/repos/svn/mozillaorg/trunk/locales/ locale
+mkdir locale
+cd locale
+svn co https://svn.mozilla.org/projects/mozilla.com/trunk/locales/ .
 
-ln -s ~/repos/svn/mozillaorg/trunk/locales/ locale
-# mkdir locale
-# cd locale
-# svn co https://svn.mozilla.org/projects/mozilla.com/trunk/locales/ .
+
+echogreen "Sync database schemas"
+./bin/sync_all
+
